+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+- Add `--override-existing-dest` option to CLI, allowing override without needing `--inplace`.
+
 ### Changed
 - Split library crate `anvil-recompress-engine` from the CLI crate `anvil-recompress`.
   - Library crate is re-exported from the CLI crate, so backwards compatibility is preserved.